refactor: moved interfaces to a separate module
